:recycle: Ground/Airborne 액션 핸들러 코드 분리
